Add negative indices test case and shift unsqueeze test under its respective dir
